cachematrix.R
Target is to write a pair of functions that cache the inverse of a matrix.
Function makeCacheMatrix: This function creates a special "matrix" object that can cache its inverse.
Function cacheSolve: This function computes the inverse of the special "matrix" returned by function makeCacheMatrix. 
If the inverse has already been calculated (and the matrix has not changed), then the cachesolve should retrieve the inverse from the cache.